Move functions out of example
