+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod prelude {
     pub use crate::source::{listener_update, Listener, SpatialAudioPlugin};
     pub use steam_audio::prelude::*;
 }
+
